python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
autogen-agentchat>=0.2.40
autogen-ext[openai]>=0.6.4
autogenstudio>=0.0.56
pinecone[grpc]>=5.0.1
openai>=1.66.5
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
websockets>=12.0
python-multipart>=0.0.6